[Linux] make renderer public
Makes the FlRenderer public so custom embedders could utilize it.
