+++ shell/platform/linux/BUILD.gn
@@ -60,6 +60,7 @@ _public_headers = [
   "public/flutter_linux/fl_pixel_buffer_texture.h",
   "public/flutter_linux/fl_plugin_registrar.h",
   "public/flutter_linux/fl_plugin_registry.h",
+  "public/flutter_linux/fl_renderer.h",
   "public/flutter_linux/fl_standard_message_codec.h",
   "public/flutter_linux/fl_standard_method_codec.h",
   "public/flutter_linux/fl_string_codec.h",
@@ -91,6 +92,7 @@ source_set("flutter_linux_sources") {
              "fl_method_call_private.h",
              "fl_method_channel_private.h",
              "fl_method_codec_private.h",
+             "fl_renderer_private.h",
              "fl_plugin_registrar_private.h",
              "fl_window_state_monitor.h",
              "key_mapping.h",
