+++ shell/platform/linux/public/flutter_linux/flutter_linux.h
@@ -24,6 +24,7 @@
 #include <flutter_linux/fl_pixel_buffer_texture.h>
 #include <flutter_linux/fl_plugin_registrar.h>
 #include <flutter_linux/fl_plugin_registry.h>
+#include <flutter_linux/fl_renderer.h>
 #include <flutter_linux/fl_standard_message_codec.h>
 #include <flutter_linux/fl_standard_method_codec.h>
 #include <flutter_linux/fl_string_codec.h>
